requirements.txt
django==3.*
djangorestframework==3.11.*
pytest==5.3.*
pytest-django==3.5.*